# perf(parser): add fast path for string starts with