not sure why theres is no diff on codespeed, locally cargo bench up to a 2.4% perf improvement in the lexer, however this may not be worth merging

```
$ cargo bench
lexer/checker.ts        time:   [5.3512 ms 5.3577 ms 5.3649 ms]
                        change: [−1.8491% −1.5567% −1.2705%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  9 (9.00%) high mild
  6 (6.00%) high severe
lexer/cal.com.tsx       time:   [2.1744 ms 2.1888 ms 2.2141 ms]
                        change: [−0.7152% +0.0449% +1.3098%] (p = 0.95 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
lexer/RadixUIAdoptionSection.jsx
                        time:   [2.1656 µs 2.1693 µs 2.1734 µs]
                        change: [−1.4098% −1.0607% −0.6859%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe
Benchmarking lexer/pdf.mjs: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.0s, enable flat sampling, or reduce sample count to 50.
lexer/pdf.mjs           time:   [1.3786 ms 1.3808 ms 1.3833 ms]
                        change: [−1.5758% −1.2052% −0.8289%] (p = 0.00 < 0.05)
                        Change within noise threshold.
lexer/antd.js           time:   [9.6186 ms 9.6321 ms 9.6465 ms]
                        change: [−2.8243% −2.3726% −1.9490%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild
```