@camc314 2025-04-25
doesn't look worth it:

```
std_starts_with:
        cmp     rsi, 3
        jb      .LBB0_1
        movzx   eax, word ptr [rdi]
        xor     eax, 28518
        movzx   ecx, byte ptr [rdi + 2]
        xor     ecx, 111
        or      cx, ax
        sete    al
        ret
.LBB0_1:
        xor     eax, eax
        ret

faster_starts_with:
        cmp     rsi, 3
        jae     .LBB1_2
        xor     eax, eax
        ret
.LBB1_2:
        movzx   eax, word ptr [rdi]
        movzx   ecx, byte ptr [rdi + 2]
        shl     ecx, 16
        or      ecx, eax
        cmp     ecx, 7303014
        sete    al
        ret
```

slightly fewer ops in the new one, but the compiler optimizes std lib pretty well